Fix bit shift wraping error in INDEX_TO_FLAG macro
* Added initial dummy codes for WorldsManager(not sure about
  architecture yet)
* Fixed bug in SlotAllocator that overwrites allocated slot in
  freeTailIdx
* Added releaseModule() function in module manager

+++ Source/Runtime/CoreObjects/Public/CoreObjectAllocator.h
@@ -31,8 +31,7 @@ namespace cbe
 class COREOBJECTS_EXPORT ObjectAllocatorBase
 {
 protected:
-    // We are not going to have same object type more than max of uint32
-    BitArray<uint32> allocValidity;
+    BitArray<uint64> allocValidity;
 
 public:
     using AllocIdx = ObjectAllocIdx;
